remove storage/shared, instead - have a top-level log file (todo: move into package) - move label definition into storage base
